pyproject.toml
[tool.poetry]
name = "vk_stickers_api"
version = "0.1.0"
description = ""
authors = ["lordralinc <46781434+lordralinc@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
aiohttp = "^3.8.1"
tortoise-orm = "^0.19.2"
uvicorn = "^0.18.2"
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
